refactor: split code for reditability
