+++ frontend/src/library-problem-creator/components/index.ts
@@ -1,5 +1,2 @@
 export { default as LibraryProblemCreator } from './LibraryProblemCreator';
 export { default as LibraryProblemCreatorResponse } from './LibraryProblemCreatorResponse';
-export { default as ProblemTypeBadge } from './ProblemTypeBadge';
-export { default as QuestionCard } from './QuestionCard';
-export { default as QuestionEditor } from './QuestionEditor';
