+++ frontend/src/library-problem-creator/index.ts
@@ -1,1 +1,10 @@
 export { LibraryProblemCreator, LibraryProblemCreatorResponse } from './components';
+export {
+  LibraryProblemCreatorProvider,
+  useLibraryProblemCreatorContext,
+} from './context/LibraryProblemCreatorContext';
+export type { LibraryProblemCreatorProviderProps } from './context/LibraryProblemCreatorContext';
+export type {
+  CreatorStep, Question, Choice, Olx,
+} from './types';
+export { PROBLEM_TYPE_LABELS, getProblemTypeLabel } from './utils/problemTypes';
